优化RunningWindow视觉效果

+++ AutumnBox.OpenFramework/Open/IExtensionUIController.cs
@@ -14,7 +14,7 @@ namespace AutumnBox.OpenFramework.Open
         /// <summary>
         /// 进度条
         /// </summary>
-        int ProgressValue { set; }
+        double ProgressValue { set; }
         /// <summary>
         /// 简要状态
         /// </summary>
